Support 'second' as period for Rate Limiting
The gateway currently supports only two periods for rate limiting: `minute` and `hour`, since rate limiting is implemented as sliding window in Skipper Ingress using a minute can be dangerous to use for high traffic applications. For example, if number of requests should be limited to 1000 requests per second, currently it needs to be set to 60,000 requests per minute, which in turns means in worst case client can make up to 60,000 requests per second.

Signed-off-by: Pawel Zaleski <pwl.zaleski@gmail.com>